+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
 - (Feature) Add Generics & Drop policy/v1beta1 support
 - (Feature) Add Kubernetes Client logger
+- (Feature) CreationFailed ArangoMember Phase
 - (Bugfix) Fix Rebalancer NPE in case if member is missing in Status
 
 ## [1.2.24](https://github.com/arangodb/kube-arangodb/tree/1.2.24) (2023-01-25)